Prevent PrismJS from modifying the document
PrismJS is not a Node module that can be used properly when needed. As
soon as it’s running in the browser, it starts to highlight everything
automatically. This is not the use case for the PIMD PrismJS plugin.
So far the only way to prevent such side-effects is defining a `Prism`
object in the browser in case PIMD is running in a browser.